Strip leading country code without leading plus
Given a country like Honduras with a country code of 504, if a number
formatted like (504) 2221-6592, it will not parse as it's an
internalation string, however it does not start with a plus or the
international prefix.

Now the territory will check if it starts with the country code before
parsing and strip the country code when it is prefixed.