Return `nil` if `map_detect` has no results
In the case that the number does not match or is invalid we don't want
to blow up in `map_detect`. We want to check for validity and return
false if it's not valid.

Removing the nil in 159a437 causes invalid numbers to throw a
`NoMethodError: undefined method `parse_national_string'` error because
when there's no match it's not returning nil.