Fix: regex error with leading_digits
When parsing `leading_digits` for number formats, `^` is being
concatenated to the format, but many leading_digit formats contain the
`|` operator. For this reason, `Format#match` returns some false
positives. I.E., /^2|5[56]/ matches the phone number 1582380560.

This commit wraps the leading_digits expression in ()s before applying
the ^ so all of the formats must match the start of the line.