Fix: do not insert an extra space in formatted numbers
When we generate the national format for a given phone number,
the output format is specified by the global_phone.json database.
For example, New Zealand is '$1-$2 $3' where $1 is the national prefix,
$2 is the first group of digits, and $3 is the last group of digits.
However, when the national prefix is added to the number, we are
erroneously inserting an extra space between the NP and the first group.
For NZ, this results in '09 -123 1234', and for other countries it results
in a double space: '02  1234 1234'.
Since the spacing is included in the format specified by the DB,
we should not need to add any additional spacing here.